+++ index.ts
@@ -1,2 +1,2 @@
-export * from './src/components/listPagination';
+export * from './src/components/listing';
 export * from './src/components/pagination';