+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 1.0.1 (Jun 7, 2024)
+
+- updated component name from "ListPagination to Listing".
+- fixed count of displayed numbers.
+- removed unnecessary properties.
+
 ## 1.0.0 (Jun 6, 2024)
 
 ### releasing